Split off KVObject to its own file, plus moved files to VRF/KeyValues
